format, and lint fix

+++ apps/nextjs/src/app/_components/PlateCard.tsx
@@ -1,10 +1,13 @@
 "use client";
 
-import { useCallback, useEffect, useState } from "react";
+import { useCallback, useState } from "react";
 import Image from "next/image";
-import { zodResolver } from "@hookform/resolvers/zod";
+// import { zodResolver } from "@hookform/resolvers/zod";
+import type { FieldValues, SubmitHandler } from "react-hook-form";
 import { useForm } from "react-hook-form";
-import { z } from "zod";
+import { /*Schema */ z } from "zod";
+// import type {Session, User, UserResponse, AuthError} from "@acme/auth";
+import type { AuthError, User } from "@acme/auth";
 
 import { Button } from "@acme/ui/button";
 import {
@@ -18,6 +21,8 @@ import { Input } from "@acme/ui/input";
 
 import UploadImg from "~/app/_components/UploadImg";
 import { createClient } from "~/utils/supabase/client";
+import type { PostgrestError, SupabaseClient } from "@supabase/supabase-js";
+// import type { GenericSchema } from "@supabase/supabase-js/dist/module/lib/types";
 
 export const PlateSchema = z.object({
   title: z.string(),
@@ -27,53 +32,59 @@ export const PlateSchema = z.object({
 });
 
 export interface Plate {
-  title: string | null;
-  descr: string | null;
-  img: string | null;
+  id?: string | null;
+  title?: string | null;
+  descr?: string | null;
+  img?: string | null;
+  id_user?: string,
 }
+
+export type ErrorSave = AuthError | null
+export type SaveEvent = (error: PostgrestError | null, data: Plate | null, plate?: Plate, errorRemoveFile?: ErrorSave) => void;
 
 export interface PropsCard {
-  user: any;
   plate: Plate;
-  onSave: (error: any, data: Plate) => never;
+  onSave?: SaveEvent;
 }
 
-async function deleteImg(supabase, plate, onSave) {
-  const _filePath = plate?.img ? plate.img.split("/").pop() : null;
+export interface PropsCardForm { user: User, onSave?: SaveEvent; }
+
+async function deleteImg(supabase: SupabaseClient, plate: Plate, onSave: SaveEvent) {
+  if (!plate.img || !plate.id) return { data: null, error: null }
+  const _filePath = plate.img ? plate.img.split("/").pop() : null;
 
   const { data, error } = await supabase
     .from("users_images")
     .delete()
     .eq("id", plate.id);
   if (error) {
-    onSave(error, null, plate);
+    onSave(error, null, plate, null);
     return { data, error };
-  } else if (_filePath) {
-    // const {data, error} =
-    const res = await supabase.storage.from("plates").remove([_filePath]);
   }
-  onSave(error, data, plate);
+  const res = _filePath && await supabase.storage.from("plates").remove([_filePath]);
+
+  onSave(error, data, plate, res as ErrorSave);
   return { data, error };
 }
 
-export function PlateCard({ user, plate, onSave }: any) {
+export function PlateCard({ plate, onSave }: PropsCard) {
   const supabase = createClient();
 
   const handleDelete = useCallback(async () => {
-    const onDelete = (error, data, plate) => onSave(error, data, plate);
+    const onDelete: SaveEvent = (error, data, plate) => onSave && onSave(error, data, plate);
     const res = await deleteImg(supabase, plate, onDelete);
     if (res.error) alert("error on delete");
-  }, [onSave, supabase]);
+  }, [onSave, plate, supabase]);
 
   return (
-    <Card size="3">
+    <Card>
       <CardTitle className="justify-center p-4 text-center text-2xl">
         <h2>{plate.title}</h2>
       </CardTitle>
       <CardContent>
         <div className="">
           <Image
-            src={plate.img}
+            src={plate.img ?? ''}
             alt="img"
             layout="fill"
             className="!relative"
@@ -96,12 +107,9 @@ export function PlateCard({ user, plate, onSave }: any) {
   );
 }
 
-export function PlateCardForm({ user, plate, onSave }: PropsCard) {
-  const [formData, setFormData] = useState({});
+export function PlateCardForm({ user }: PropsCardForm) {
   const [loading, setLoading] = useState(true);
-  const [imgUrl, setImgUrl] = useState(
-    plate && plate.img && plate.img,
-  );
+  const [imgUrl, setImgUrl] = useState<string | null>(null);
 
   const supabase = createClient();
 
@@ -110,52 +118,53 @@ export function PlateCardForm({ user, plate, onSave }: PropsCard) {
     register,
     handleSubmit,
     reset,
-    formState: { errors, isSubmitting, isValid },
-    defaultValues,
+    formState: { /* errors , */ isSubmitting, isValid },
+    /* defaultValues ,*/
   } = form;
 
   const updatePlate = useCallback(
-    async ({ title, descr, img, imgUrl }: Plate) => {
+    async ({ title, descr }: Plate) => {
       // console.log('onSubmit data', fullname, );
       try {
         setLoading(true);
-
-        const { data, error } = await supabase.from("users_images").upsert({
-          id_user: user?.id as string,
+        const dataForm: Plate = {
+          id_user: user.id,
           title,
           descr,
-          img: imgUrl || "",
+          img: imgUrl ?? "",
           // updated_at: new Date().toISOString(),
-        });
+        }
+
+        const { error } = await supabase.from("users_images").upsert(dataForm);
 
         if (error) {
           console.log("error update plate", error);
-          alert("Error on update" + error);
+          alert("Error on update" + error.message);
           return false;
         }
         setImgUrl(null);
         reset();
-        onSave && onSave(data);
+        // onSave(error, dataForm, dataForm );
         alert("plate updated!");
       } catch (error) {
-        alert("Error updating the data!", error);
         setLoading(false);
-        alert("Error on update" + error);
+        console.log('PalteCard error on update form', error)
+        alert("Error on update");
       } finally {
         setLoading(false);
       }
     },
-    [setLoading, user, setImgUrl, onSave],
+    [user.id, imgUrl, supabase, reset],
   );
 
-  function onSubmit(dataForm) {
-    updatePlate({ ...dataForm, imgUrl });
+  const onSubmit: SubmitHandler<FieldValues> = (dataForm) => {
+    updatePlate({ ...dataForm as unknown as Plate, img: imgUrl }).then(() => null).catch(() => null);
     console.log("onSubmit data", dataForm);
   }
 
   return (
     <form onSubmit={handleSubmit(onSubmit)} action="/" className="mt-8">
-      <Card size="3">
+      <Card>
         <CardTitle className="justify-center p-4 text-center text-2xl">
           <div className="mt-2">
             <Input
@@ -170,7 +179,7 @@ export function PlateCardForm({ user, plate, onSave }: PropsCard) {
         <CardContent>
           <div className="">
             <UploadImg
-              uid={user?.id ?? null}
+              uid={user.id}
               url={imgUrl}
               size={150}
               bucket="plates"
@@ -197,11 +206,10 @@ export function PlateCardForm({ user, plate, onSave }: PropsCard) {
         <Button
           type="submit"
           disabled={!isValid}
-          className={`${
-            !isValid ? "bg-slate-400 hover:bg-slate-400" : undefined
-          } inline-flex w-full items-center justify-center rounded-md bg-black px-3.5 py-2.5 font-semibold leading-7 text-white hover:bg-black/80`}
+          className={`${!isValid ? "bg-slate-400 hover:bg-slate-400" : undefined
+            } inline-flex w-full items-center justify-center rounded-md bg-black px-3.5 py-2.5 font-semibold leading-7 text-white hover:bg-black/80`}
         >
-          {isSubmitting ? "Submitting..." : "Save Plate"}
+          {isSubmitting && loading ? "Submitting..." : "Save Plate"}
         </Button>
       </div>
     </form>
